Export HDF5 files via ArrayData from mkstd
