--- petab_sciml/hdf5/__init__.py
@@ -1,1 +1,0 @@
-from .write_hdf5 import *
